api_funcs/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic==2.4.2
python-jose[cryptography]==3.3.0
email-validator==2.1.0
httpx==0.25.1
python-multipart==0.0.6
